+++ src/FlaUILibrary/keywords/radiobutton.py
@@ -1,6 +1,7 @@
 from robotlibcore import keyword
 from FlaUILibrary.flaui.interface import InterfaceType
 from FlaUILibrary.flaui.module import ToggleButton
+from FlaUILibrary.flaui.uia import UIA
 
 
 class RadioButtonKeywords:
@@ -8,7 +9,7 @@ class RadioButtonKeywords:
     Interface implementation from robotframework usage for radio button keywords.
     """
 
-    def __init__(self, module):
+    def __init__(self, module: UIA):
         """
         Constructor for radiobutton keywords.