+++ src/FlaUILibrary/keywords/tab.py
@@ -1,6 +1,7 @@
 from robotlibcore import keyword
 from FlaUILibrary.flaui.interface import InterfaceType
 from FlaUILibrary.flaui.module import Tab
+from FlaUILibrary.flaui.uia import UIA
 
 
 class TabKeywords:
@@ -8,7 +9,7 @@ class TabKeywords:
     Interface implementation from robotframework usage for tab keywords.
     """
 
-    def __init__(self, module):
+    def __init__(self, module: UIA):
         """
         Constructor for tab keywords.